booking a booth ALMOST works - lightly tested(needs more)
booking a booth is very close to working. only problem is that after
booking a booth, it does not show the booth as red right away, takes a
refresh of the page. not sure why this is happening, i have a function
call that SHOULD work, but it is not. figure it out later.
also considering putting a timer around refreshes so that they always
see an up to date schedule