DONE!, 1 bug to think about
BUG: when you book a booth, an alert comes up saying congrats, but the
page reloads shortly after the alert comes up, not sure how to make the
refresh wait till user presses ok on alert
can now book and cancel and banned message shows up when they try to
book booth.
page is reloaded, but saving the date in 'session storage' (it will only
last as long as that tab is open). so its reloaded but stays on same
date